--- a/recommendation/pytorch/convert.py
+++ b/recommendation/pytorch/convert.py
@@ -4,8 +4,12 @@ from collections import defaultdict
 
 import numpy as np
 import pandas as pd
+from tqdm import tqdm
 
 from load import implicit_load
+
+
+MIN_RATINGS = 20
 
 
 USER_COLUMN = 'user_id'
@@ -35,10 +39,13 @@ def main():
     args = parse_args()
     np.random.seed(args.seed)
 
+    print("Loading raw data from {}".format(args.path))
     df = implicit_load(args.path, sort=False)
+    print("Filtering out users with less than {} ratings".format(MIN_RATINGS))
     grouped = df.groupby(USER_COLUMN)
-    df = grouped.filter(lambda x: len(x) >= 20)
+    df = grouped.filter(lambda x: len(x) >= MIN_RATINGS)
 
+    print("Mapping original user and item IDs to new sequential IDs")
     original_users = df[USER_COLUMN].unique()
     original_items = df[ITEM_COLUMN].unique()
 
@@ -51,17 +58,20 @@ def main():
     assert df[USER_COLUMN].max() == len(original_users) - 1
     assert df[ITEM_COLUMN].max() == len(original_items) - 1
 
+    print("Creating list of items for each user")
     # Need to sort before popping to get last item
     df.sort_values(by='timestamp', inplace=True)
     all_ratings = set(zip(df[USER_COLUMN], df[ITEM_COLUMN]))
     user_to_items = defaultdict(list)
-    for row in df.itertuples():
+    for row in tqdm(df.itertuples(), desc='Ratings', total=len(df)):
         user_to_items[getattr(row, USER_COLUMN)].append(getattr(row, ITEM_COLUMN))  # noqa: E501
 
     test_ratings = []
     test_negs = []
     all_items = set(range(len(original_items)))
-    for user in range(len(original_users)):
+    print("Generating {} negative samples for each user"
+          .format(args.negatives))
+    for user in tqdm(range(len(original_users)), desc='Users', total=len(original_users)):  # noqa: E501
         test_item = user_to_items[user].pop()
 
         all_ratings.remove((user, test_item))
@@ -71,7 +81,7 @@ def main():
         test_ratings.append((user, test_item))
         test_negs.append(list(np.random.choice(all_negs, args.negatives)))
 
-    # serialize
+    print("Saving train and test CSV files to {}".format(args.output))
     df_train_ratings = pd.DataFrame(list(all_ratings))
     df_train_ratings['fake_rating'] = 1
     df_train_ratings.to_csv(os.path.join(args.output, TRAIN_RATINGS_FILENAME),