--- a/recommendation/pytorch/run_and_time.sh
+++ b/recommendation/pytorch/run_and_time.sh
@@ -3,7 +3,7 @@
 # to use the script:
 #   run_and_time.sh <random seed 1-5>
 
-THRESHOLD=0.9562
+THRESHOLD=0.6289
 BASEDIR=$(dirname -- "$0")
 
 # start timing
@@ -19,15 +19,15 @@ if unzip ml-20m.zip
 then
     echo "Start processing ml-20m/ratings.csv"
     t0=$(date +%s)
-	python $BASEDIR/convert.py ml-20m/ratings.csv ml-20m
+	python $BASEDIR/convert.py ml-20m/ratings.csv ml-20m --negatives 999
     t1=$(date +%s)
 	delta=$(( $t1 - $t0 ))
     echo "Finish processing ml-20m/ratings.csv in $delta seconds"
 
     echo "Start training"
     t0=$(date +%s)
-	python $BASEDIR/ncf.py ml-20m -l 0.0005 -b 16384 --layers 256 128 64 -f 64 \
-		--seed $seed --threshold $THRESHOLD
+	python $BASEDIR/ncf.py ml-20m -l 0.0005 -b 2048 --layers 256 128 64 -f 64 \
+		--seed $seed --threshold $THRESHOLD --processes 10
     t1=$(date +%s)
 	delta=$(( $t1 - $t0 ))
     echo "Finish training in $delta seconds"