This pull request addresses [mlperf/policies#14](https://github.com/mlperf/policies/issues/14) by increasing the number of negative samples from 100 to 999 for the recommendation benchmark and updating the threshold accordingly. The new threshold is a hit rate of 0.6289 at 10, which is based on the lowest hit rate achieved by the 5 random seeds:

```
seed
1    0.629786
2    0.633101
3    0.632241
4    0.628985
5    0.630992
Name: hit_rate, dtype: float64
```

[Here](https://github.com/mlperf/reference/files/2112584/n999_timing_output.txt) is the output using the new threshold with the 5 seeds.

Calculating validation accuracy was also improved to parallelize across CPUs, decreasing the validation time per epoch from ~90 seconds to ~53 seconds. Batch size also changed based on hyperparameter tuning from 16384 to 2048. This increased per epoch training time from ~190 seconds to ~350 seconds. Total end-to-end training time now ranged from 3960 seconds to 4772 with a median, mean, and standard deviation of 4361, 4338.4, and 261.74 respectively.